+++ src/v3/selector.rs
@@ -241,6 +241,7 @@ where
             io.clone(),
             mqtt::Codec::default().max_size(self.max_size),
             16,
+            false,
             self.pool.clone(),
         ));
         let mut timeout = Deadline::new(self.handshake_timeout);
@@ -318,6 +319,7 @@ where
             io.get_ref(),
             mqtt::Codec::default().max_size(self.max_size),
             16,
+            false,
             self.pool.clone(),
         ));